[CodeGen] add string abi, add memcpy builtin function

+++ test/release/0-0-1.sol
@@ -1,38 +1,38 @@
 pragma solidity ^0.5.0;
 
 contract SafeMath {
-    function add(uint64 a, uint64 b) public pure returns (uint64) {
-        uint64 c = a + b;
+    function add(uint256 a, uint256 b) public pure returns (uint256) {
+        uint256 c = a + b;
         require(c >= a, "SafeMath: addition overflow");
 
         return c;
     }
 
-    function sub(uint64 a, uint64 b) public pure returns (uint64) {
+    function sub(uint256 a, uint256 b) public pure returns (uint256) {
         require(b <= a, "SafeMath: subtraction overflow");
-        uint64 c = a - b;
+        uint256 c = a - b;
 
         return c;
     }
 
-    function mul(uint64 a, uint64 b) public pure returns (uint64) {
+    function mul(uint256 a, uint256 b) public pure returns (uint256) {
         if (a == 0) {
             return uint64(0);
         }
 
-        uint64 c = a * b;
+        uint256 c = a * b;
         require(c / a == b, "SafeMath: multiplication overflow");
 
         return c;
     }
 
-    function div(uint64 a, uint64 b) public pure returns (uint64) {
+    function div(uint256 a, uint256 b) public pure returns (uint256) {
         require(b > 0, "SafeMath: division by zero");
-        uint64 c = a / b;
+        uint256 c = a / b;
         return c;
     }
 
-    function mod(uint64 a, uint64 b) public pure returns (uint64) {
+    function mod(uint256 a, uint256 b) public pure returns (uint256) {
         require(b != 0, "SafeMath: modulo by zero");
         return a % b;
     }
